pyproject.toml
[tool.black]
skip-string-normalization = true  # Optional: Prevent string quotes normalization if desired

[tool.ruff]
exclude = ["tests/"]  # Exclude specific directories

[tool.ruff.lint]
ignore = ["E501", "ANN", "D107", "S603", "D100", "D101", "E402", "E501"]  # E501: Ignore line length errors globally; ANN: Ignore untyped annotation warnings; D107: Ignore missing docstring in __init__; S603: Ignore subprocess call warning; D100: Ignore missing docstring in public module; D101: Ignore missing docstring in public class; E402: Ignore module level import not at top of file
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "C90", # mccabe
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "A",   # flake8-builtins
]


[tool.ruff.lint.pydocstyle]
convention = "google"  # Enforce Google-style docstrings

[tool.ruff.lint.mccabe]
max-complexity = 10  # Limit cyclomatic complexity

[tool.pytest.ini_options]
testpaths = ["tests"]  # Define test directory
pythonpath = ["."]   # Add current directory to Python path

[project]
dependencies = [
    "fuzzywuzzy",
    "python-Levenshtein",  # Optional but recommended for better performance
    "fastapi",
    "pydantic",
    "langchain",
    "langchain-openai",
    "pyyaml",
    "pytest",
    "pytest-mock",
    "python-dotenv",
    "uvicorn",
    "requests"
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true  # Avoid errors on missing imports
disallow_untyped_defs = true   # Enforce typing in non-test code
strict_optional = true         # Enable strict optional checking
exclude = "tests"              # Exclude tests directory from mypy checks
